requirements.txt
django
django-crispy-forms
python-slugify